# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StoryInput {
  /** Story category (e.g. AITA, relationship drama, etc.) */
  category: string;
  /**
     * Optional custom post title to use
     * @nullable
     */
  customTitle?: string | null;
}

export interface StoryResponse {
  /** Reddit post title */
  title: string;
  /** Full story text for narration */
  story: string;
}

export interface Voice {
  /** Voice ID */
  id: string;
  /** Voice name */
  name: string;
  /** Voice category (e.g. premade, cloned) */
  category: string;
  /**
     * URL to preview audio clip
     * @nullable
     */
  previewUrl: string | null;
}

export interface SpeakInput {
  /** Text to convert to speech */
  text: string;
  /** ElevenLabs voice ID to use */
  voiceId: string;
}

export interface WordTimestamp {
  /** The word */
  word: string;
  /** Start time in seconds */
  start: number;
  /** End time in seconds */
  end: number;
}

export interface SpeakResponse {
  /** Base64-encoded MP3 audio */
  audioBase64: string;
  /** Word-level timestamps for caption sync */
  words: WordTimestamp[];
}

export interface ErrorResponse {
  error: string;
}