# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: story
    description: Reddit story generation
  - name: tts
    description: Text to speech
  - name: voices
    description: Voice management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /story/generate:
    post:
      operationId: generateStory
      tags: [story]
      summary: Generate a Reddit-style story
      description: Uses OpenRouter AI to generate a Reddit-style AITA or drama story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryInput"
      responses:
        "200":
          description: Generated story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryResponse"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /voices:
    get:
      operationId: listVoices
      tags: [voices]
      summary: List available TTS voices
      description: Returns all available ElevenLabs voices
      responses:
        "200":
          description: List of voices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voice"
        "500":
          description: Failed to fetch voices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tts/speak:
    post:
      operationId: speakText
      tags: [tts]
      summary: Convert text to speech with word timestamps
      description: Uses ElevenLabs to generate audio and returns base64 audio plus word-level timestamps for caption sync
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakInput"
      responses:
        "200":
          description: Audio and timestamps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakResponse"
        "500":
          description: TTS failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    StoryInput:
      type: object
      required: [category]
      properties:
        category:
          type: string
          description: Story category (e.g. AITA, relationship drama, etc.)
        customTitle:
          type: ["string", "null"]
          description: Optional custom post title to use
    StoryResponse:
      type: object
      required: [title, story]
      properties:
        title:
          type: string
          description: Reddit post title
        story:
          type: string
          description: Full story text for narration
    Voice:
      type: object
      required: [id, name, category, previewUrl]
      properties:
        id:
          type: string
          description: Voice ID
        name:
          type: string
          description: Voice name
        category:
          type: string
          description: Voice category (e.g. premade, cloned)
        previewUrl:
          type: ["string", "null"]
          description: URL to preview audio clip
    SpeakInput:
      type: object
      required: [text, voiceId]
      properties:
        text:
          type: string
          description: Text to convert to speech
        voiceId:
          type: string
          description: ElevenLabs voice ID to use
    WordTimestamp:
      type: object
      required: [word, start, end]
      properties:
        word:
          type: string
          description: The word
        start:
          type: number
          description: Start time in seconds
        end:
          type: number
          description: End time in seconds
    SpeakResponse:
      type: object
      required: [audioBase64, words]
      properties:
        audioBase64:
          type: string
          description: Base64-encoded MP3 audio
        words:
          type: array
          items:
            $ref: "#/components/schemas/WordTimestamp"
          description: Word-level timestamps for caption sync
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string