# lib/api-zod/src/generated/types/speakInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SpeakInput {
  /** Text to convert to speech */
  text: string;
  /** ElevenLabs voice ID to use */
  voiceId: string;
}