# lib/api-zod/src/generated/types/speakResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WordTimestamp } from './wordTimestamp';

export interface SpeakResponse {
  /** Base64-encoded MP3 audio */
  audioBase64: string;
  /** Word-level timestamps for caption sync */
  words: WordTimestamp[];
}