# lib/api-zod/src/generated/types/storyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StoryInput {
  /** Story category (e.g. AITA, relationship drama, etc.) */
  category: string;
  /**
     * Optional custom post title to use
     * @nullable
     */
  customTitle?: string | null;
}