# lib/api-zod/src/generated/types/voice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Voice {
  /** Voice ID */
  id: string;
  /** Voice name */
  name: string;
  /** Voice category (e.g. premade, cloned) */
  category: string;
  /**
     * URL to preview audio clip
     * @nullable
     */
  previewUrl: string | null;
}